pyproject.toml
[tool.poetry]
name = "openai-proxy"
version = "0.1.0"
description = ""
authors = ["celeste <parafactual@gmail.com>"]
readme = "README.md"
packages = [{include = "openai_proxy"}]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.3"
openai = "^0.27.2"
python-dotenv = "^1.0.0"
gunicorn = "^20.1.0"
flask-cors = "^3.0.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
typer = "^0.9.0"
peewee = "^3.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ipdb = "^0.13.13"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
